services/pricing-service/requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
redis==4.5.5